requirements.txt
click~=8.1.3
setuptools~=63.2.0
cassandra-driver~=3.25.0